requirements.txt
pandas
numpy
fastapi
sseclient-py
urllib3
tiktoken
